Add Get properties in the file GameManager to be used in the Exp3 file.
Add changes in the GameState file to include the piece info when the user put it in the right spot. In the other State files added an IF condition to detect when the piece info is null and change it for the new variable added in the Therapist file and updated in GameState.